update variable  for age65&above estimate (count)
originally used same var for estimate count and percentage (from S0101), then realized count and percentage could be from different table, so look further to locate estimate count in S0103. BUT this table is not as complete as S0101, a lot of NAs in the estimate count. So considering using percentage * total population to calculate count (need rounding up).